Redesign traces
